REFACT: Split unit tests - split `tests/test_information_package.py` into archive and struct handling parts.
